requirements.txt
requests
pandas
numpy
python-dotenv
beautifulsoup4
lxml